@@ -3,7 +3,7 @@
  *
  *   This file implements the file IO of the Picross solver
  *
- * Copyright (c) 2010-2020 Pierre DEJOUE
+ * Copyright (c) 2010-2022 Pierre DEJOUE
  ******************************************************************************/
 #include <picross/picross_io.h>
 
@@ -170,7 +170,7 @@ private:
         return "UNKNOWN";
     }
 
-    void error_decorator(const ErrorHandler& error_handler, const std::string& msg)
+    void error_decorator(const ErrorHandler& error_handler, const std::string_view& msg)
     {
         std::ostringstream oss;
         oss << msg << " (parsing_state = " << parsing_state_str() << ")";
@@ -381,13 +381,13 @@ private:
 
 
 template <typename F>
-std::vector<InputGrid> parse_input_file_generic(const std::string& filepath, const ErrorHandler& error_handler) noexcept
+std::vector<InputGrid> parse_input_file_generic(std::string_view filepath, const ErrorHandler& error_handler) noexcept
 {
     std::vector<InputGrid> result;
 
     try
     {
-        std::ifstream inputstream(filepath);
+        std::ifstream inputstream(filepath.data());
         if (inputstream.is_open())
         {
             FileParser<F> parser;
@@ -399,7 +399,7 @@ std::vector<InputGrid> parse_input_file_generic(const std::string& filepath, con
             {
                 line_nb++;
                 inputstream.getline(line, INPUT_BUFFER_SZ - 1);
-                parser.parse_line(std::string(line), result, [line_nb, &line, &error_handler](const std::string& msg, ExitCode code)
+                parser.parse_line(std::string(line), result, [line_nb, &line, &error_handler](std::string_view msg, ExitCode code)
                     {
                         std::ostringstream oss;
                         oss << "Parsing error [" << msg << "] on line " << line_nb << ": " << line;
@@ -457,12 +457,12 @@ inline void write_metadata_non_format(std::ostream& ostream, const std::map<std:
 
 } // anonymous namespace
 
-std::vector<InputGrid> parse_input_file(const std::string& filepath, const ErrorHandler& error_handler) noexcept
+std::vector<InputGrid> parse_input_file(std::string_view filepath, const ErrorHandler& error_handler) noexcept
 {
     return parse_input_file_generic<FileFormat::Native>(filepath, error_handler);
 }
 
-std::vector<InputGrid> parse_input_file_non_format(const std::string& filepath, const ErrorHandler& error_handler) noexcept
+std::vector<InputGrid> parse_input_file_non_format(std::string_view filepath, const ErrorHandler& error_handler) noexcept
 {
     return parse_input_file_generic<FileFormat::Non>(filepath, error_handler);
 }
